refactor: extract address to point conversion to component
